Migrate badge component to TS
