+++ packages/bpk-component-badge/src/themeAttributes.tsx
